requirements.txt
certifi
chardet
defusedxml
django-allauth
django-ckeditor
django-js-asset
django-model-utils
django-mptt
django-notifications-hq
django-password-reset
django-taggit
idna
jsonfield
Markdown
oauthlib
Pillow
Pygments
python3-openid
pytz
requests
requests-oauthlib
urllib3